% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_image.R
\name{censor_image}
\alias{censor_image}
\title{Censor image using blurring or solid rectangle}
\usage{
censor_image(
  image,
  filename = NULL,
  path = NULL,
  automatic = TRUE,
  blur = 75,
  width = 100,
  height = 100,
  x = 100,
  y = 100,
  color = "black"
)
}
\arguments{
\item{image}{The image to be modified}

\item{filename}{The filename for the censored output image}

\item{path}{The path to the output image}

\item{automatic}{Whether to use automatic face detection (`automatic = TRUE` is default)}

\item{blur}{The amount of blur if `automatic = TRUE` (0 to 100)}

\item{width}{The width of the rectangle (only for `automatic=FALSE`)}

\item{height}{The height of the rectangle (only for `automatic=FALSE`)}

\item{x}{The x coordinate of the rectangle (only for `automatic=FALSE`)}

\item{y}{The y coordinate of the rectangle (only for `automatic=FALSE`)}

\item{color}{The color of the rectangle (only for `automatic=FALSE`)}
}
\value{
The path to the censored image file
}
\description{
This function inputs an image and applies blurring of detected faces
or solid rectangle censoring of a manually defined region.
}
