% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_video.R
\name{censor_video}
\alias{censor_video}
\title{Censor video}
\usage{
censor_video(
  video,
  filename = NULL,
  path = NULL,
  automatic = TRUE,
  rescale_rect = 1,
  width = 100,
  height = 100,
  x = 100,
  y = 100,
  color = "black"
)
}
\arguments{
\item{video}{The video to be modified}

\item{filename}{The filename for the censored output video}

\item{path}{The path to the output video}

\item{automatic}{Whether to attempt automatic face detection (`automatic = TRUE` is default)}

\item{rescale_rect}{Rescale rectangle by factor if `automatic = TRUE` (default = 1)}

\item{width}{The width of the rectangle (only for `automatic = FALSE`)}

\item{height}{The height of the rectangle (only for `automatic = FALSE`)}

\item{x}{The x coordinate of the rectangle (only for `automatic = FALSE`)}

\item{y}{The y coordinate of the rectangle (only for `automatic = FALSE`)}

\item{color}{The color of the rectangle}
}
\value{
The path to the censored video file
}
\description{
This function inputs a video and censors a manually defined region.
}
