% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_source.R
\name{cite_source}
\alias{cite_source}
\title{Cite source}
\usage{
cite_source()
}
\value{
The preferred citation format for the package and language resources
}
\description{
This function outputs the preferred citation for the
signglossR package and its main dictionary sources.
}
