% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_image.R
\name{compose_image}
\alias{compose_image}
\title{Compose a modified/annotated image from an image file}
\usage{
compose_image(
  image,
  filename = NULL,
  path = NULL,
  scale = 1,
  crop_region = NULL,
  text = NULL,
  textcolor = "white",
  boxcolor = NULL,
  font = "Helvetica",
  fontsize = 70,
  rel_fontsize = NULL,
  gravity = "southwest",
  location = "+20+20",
  border = FALSE
)
}
\arguments{
\item{image}{The path to the image file to be modified}

\item{filename}{The filename of the output image}

\item{path}{The path to the output image}

\item{scale}{The resizing proportion of the original image (0 to 1; default is 1 = original size)}

\item{crop_region}{The region to be cropped (e.g. '100x150+100+100', (width x height +upper_x +upper_y)).}

\item{text}{The text string to be added to the image}

\item{textcolor}{The color of the text annotation (default is "white")}

\item{boxcolor}{The color of the textbox (default is `NULL`)}

\item{font}{The font of the text annotation (default is "Helvetica")}

\item{fontsize}{The fontsize of the text annotation (default is 70)}

\item{rel_fontsize}{Whether to adjust the fontsize relative to width (x) or height (y)}

\item{gravity}{The anchor point of the added text (default is "southwest")}

\item{location}{The location of the text relative to the anchor point}

\item{border}{Set to `TRUE` if you want a black border frame for the output image}
}
\value{
The filename of the image file that was created
}
\description{
This function inputs an image file and allows for resizing, cropping, and text annotations
}
