% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gif.R
\name{get_gif}
\alias{get_gif}
\title{Get STS sign GIF (Svenskt teckenspråkslexikon)}
\usage{
get_gif(id, filename = NULL, path = NULL)
}
\arguments{
\item{id}{ID number for the STS sign}

\item{filename}{The filename of the downloaded image (GIF)}

\item{path}{The path to the downloaded image (GIF)}
}
\value{
The path of the image file that was downloaded
}
\description{
This function inputs an STS ID number and downloads the GIF
of the corresponding sign entry in Svenskt teckenspråkslexikon.
}
