% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gif_sts.R
\name{get_gif_sts}
\alias{get_gif_sts}
\title{Get STS sign GIF (Svenskt teckenspråkslexikon)}
\usage{
get_gif_sts(id, destination = "./")
}
\arguments{
\item{id}{ID number for the STS sign}

\item{destination}{The path/filename of the downloaded image}
}
\value{
The path of the image file that was downloaded
}
\description{
This function inputs an STS ID number and downloads the GIF
of the corresponding sign entry in Svenskt teckenspråkslexikon.
}
