% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image.R
\name{get_image}
\alias{get_image}
\title{Get sign image}
\usage{
get_image(
  id,
  destination = "./",
  acronym = "sts",
  glosstext = FALSE,
  gravity = "north",
  location = "+10+20",
  fontsize = 70,
  overlay = FALSE,
  trim = 1
)
}
\arguments{
\item{id}{ID number for the sign}

\item{destination}{The path/filename of the downloaded image}

\item{acronym}{The acronym for the target sign language (e.g. "STS")}

\item{glosstext}{Prints the ID gloss onto the image if set to `TRUE` (default = `FALSE`)}

\item{gravity}{The anchor point of the text annotation if selected (default = "north")}

\item{location}{The exact location of the text annotation relative to `gravity`}

\item{fontsize}{The fontsize for the text annotation (default = 70)}

\item{overlay}{Optional argument if overlay image wanted (defaults to FALSE)}

\item{trim}{Optional argument if horizontal trim wanted (values 0 to 1; default=1)}
}
\value{
The path of the image file that was downloaded
}
\description{
This function inputs an  ID number and downloads the
image of the corresponding sign entry in the selected language.
}
\details{
Arguments available for ASL:
A few esthetics arguments for the text annotation are optional (e.g. `gravity`, `fontsize`)

Arguments available for STS:
Options include creating side-by-side or overlay images (defaults to FALSE) and
whether a horizontal trim (0 to 1; default=1) should be applied.
}
