% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image_asl.R
\name{get_image_asl}
\alias{get_image_asl}
\title{Get ASL sign image (ASL Signbank)}
\usage{
get_image_asl(
  id,
  destination = "./",
  glosstext = FALSE,
  gravity = "north",
  location = "+10+20",
  fontsize = 70
)
}
\arguments{
\item{id}{ID number for the ASL sign}

\item{destination}{The path/filename of the downloaded image}

\item{glosstext}{Prints the ID gloss onto the image if set to `TRUE` (default = `FALSE`)}

\item{gravity}{The anchor point of the text annotation if selected (default = "north")}

\item{location}{The exact location of the text annotation relative to `gravity`}

\item{fontsize}{The fontsize for the text annotation (default = 70)}
}
\value{
The path of the image file that was downloaded
}
\description{
This function inputs an ASL ID gloss or number and downloads the
image of the corresponding sign entry in ASL Signbank.
}
\details{
A few esthetics arguments for the text annotation are optional (e.g. `gravity`, `fontsize`)
}
