% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image_sts.R
\name{get_image_sts}
\alias{get_image_sts}
\title{Get STS sign image (Svenskt teckenspråkslexikon)}
\usage{
get_image_sts(id, destination = "./", overlay = FALSE, trim = 1)
}
\arguments{
\item{id}{ID number for the STS sign}

\item{destination}{The path/filename of the downloaded image}

\item{overlay}{Optional argument if overlay image wanted (defaults to FALSE)}

\item{trim}{Optional argument if horizontal trim wanted (values 0 to 1; default=1)}
}
\value{
The path of the image file that was downloaded
}
\description{
This function inputs an STS ID number and downloads the image
of the corresponding sign entry in Svenskt teckenspråkslexikon.
}
\details{
Options include creating side-by-side or overlay images (defaults to FALSE) and
whether a horizontal trim (0 to 1; default=1) should be applied.
}
