% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{get_images}
\alias{get_images}
\title{Get multiple sign images}
\usage{
get_images(
  ids,
  directory = "./",
  acronym = "sts",
  glosstext = FALSE,
  gravity = "north",
  location = "+20+20",
  fontsize = 70,
  overlay = FALSE,
  trim = 1
)
}
\arguments{
\item{ids}{ID numbers for the signs}

\item{directory}{The destination directory for the downloaded images}

\item{acronym}{The acronym for the target sign language (e.g. "STS")}

\item{glosstext}{Prints the ID gloss onto the image if set to `TRUE` (default = `FALSE`)}

\item{gravity}{The anchor point of the text annotation if selected (default = "north")}

\item{location}{The exact location of the text annotation relative to `gravity`}

\item{fontsize}{The fontsize for the text annotation (default = 70)}

\item{overlay}{Optional argument if overlay image wanted (defaults to FALSE)}

\item{trim}{Optional argument if horizontal trim wanted (values 0 to 1; default=1)}
}
\value{
The path of the image files that were downloaded
}
\description{
This function inputs a string/vector of ID numbers and downloads the
images of the corresponding sign entries in the selected language.
}
\details{
Arguments available for ASL:
A few esthetics arguments for the text annotation are optional (e.g. `gravity`, `fontsize`)

Arguments available for STS:
Options include creating side-by-side or overlay images (defaults to FALSE) and
whether a horizontal trim (0 to 1; default=1) should be applied.
}
