% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video.R
\name{get_video}
\alias{get_video}
\title{Get sign video}
\usage{
get_video(id, acronym = "sts")
}
\arguments{
\item{id}{ID number for the sign}

\item{acronym}{The acronym for the target sign language (e.g. "STS")}
}
\value{
The name of the video file that was downloaded
}
\description{
This function inputs an ID number and a sign language acronym and
downloads the video of the corresponding sign entry in the linked resource.
}
