% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video_asl.R
\name{get_video_asl}
\alias{get_video_asl}
\title{Get ASL sign video (ASL Signbank)}
\usage{
get_video_asl(id)
}
\arguments{
\item{id}{ID number for the ASL sign}
}
\value{
The name of the video file that was downloaded
}
\description{
This function inputs an ASL ID gloss or number and downloads the video
of the corresponding sign entry in ASL Signbank.
}
