% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video_sts.R
\name{get_video_sts}
\alias{get_video_sts}
\title{Get STS sign video (Svenskt teckenspråkslexikon)}
\usage{
get_video_sts(id, destination = "./")
}
\arguments{
\item{id}{ID number for the STS sign}

\item{destination}{The path/filename of the downloaded video}
}
\value{
The name of the video file that was downloaded
}
\description{
This function inputs an STS ID number and downloads the video
of the corresponding sign entry in Svenskt teckenspråkslexikon.
}
