% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_videos.R
\name{get_videos}
\alias{get_videos}
\title{Get multiple sign videos}
\usage{
get_videos(ids, directory = "./", acronym = "sts")
}
\arguments{
\item{ids}{ID numbers for the signs}

\item{directory}{The destination directory for the downloaded videos}

\item{acronym}{The acronym for the target sign language (e.g. "STS")}
}
\value{
The path of the video files that were downloaded
}
\description{
This function inputs a string/vector of ID numbers and downloads the
videos of the corresponding sign entries in the selected language.
}
