% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss2id.R, R/gloss2id1.R
\name{gloss2id}
\alias{gloss2id}
\title{Find ID number from ID gloss}
\usage{
gloss2id(gloss, acronym = "sts")

gloss2id(gloss, acronym = "sts")
}
\arguments{
\item{gloss}{ID gloss for the sign}

\item{acronym}{The acronym for the target sign language (e.g. "STS")}
}
\value{
The unique ID gloss linked to the sign used as its label in Signbank

The unique ID gloss linked to the sign used as its label in Signbank
}
\description{
This function inputs an ID gloss and outputs an
ID number linked to the same sign (its dictionary entry).

This function inputs an ID gloss and outputs an
ID number linked to the same sign (its dictionary entry).
}
