% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss2id_asl.R
\name{gloss2id_asl}
\alias{gloss2id_asl}
\title{Find ID number from ID gloss (ASL Signbank)}
\usage{
gloss2id_asl(gloss)
}
\arguments{
\item{gloss}{ID gloss for the ASL sign}
}
\value{
The unique ID number linked to the sign used for indexing online
}
\description{
This function inputs an ASL ID gloss as a string and outputs the unique
ID number linked to the same sign (its Signbank index).
}
