% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss2id_sts.R
\name{gloss2id_sts}
\alias{gloss2id_sts}
\title{Find ID number from ID gloss (STS)}
\usage{
gloss2id_sts(gloss)
}
\arguments{
\item{gloss}{ID gloss for the sign}
}
\value{
An ID number linked to the sign gloss
}
\description{
This function inputs an STS gloss as a string and outputs the unique
ID number linked to the same sign (its Signbank index).
}
