% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id2gloss.R, R/id2gloss1.R
\name{id2gloss}
\alias{id2gloss}
\title{Find ID gloss from ID number}
\usage{
id2gloss(id, acronym = "sts")

id2gloss(id, acronym = "sts")
}
\arguments{
\item{id}{ID number for the sign}

\item{acronym}{The acronym for the target sign language (e.g. "STS")}
}
\value{
The unique ID gloss linked to the sign used as its label in Signbank

The unique ID gloss linked to the sign used as its label in Signbank
}
\description{
This function inputs an ID number as an integer or string and outputs the
unique ID gloss linked to the same sign (its dictionary entry).

This function inputs an ID number as an integer or string and outputs the
unique ID gloss linked to the same sign (its dictionary entry).
}
