% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id2gloss1.R
\name{id2gloss1}
\alias{id2gloss1}
\title{Find ID gloss from ID number}
\usage{
id2gloss1(id, acronym = "sts")
}
\arguments{
\item{id}{ID number for the sign}

\item{acronym}{The acronym for the target sign language (e.g. "STS")}
}
\value{
The unique ID gloss linked to the sign used as its label in Signbank
}
\description{
This function inputs an ID number as an integer or string and outputs the
unique ID gloss linked to the same sign (its dictionary entry).
}
