% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id2gloss_asl.R
\name{id2gloss_asl}
\alias{id2gloss_asl}
\title{Find ID gloss from ID number (ASL Signbank)}
\usage{
id2gloss_asl(id)
}
\arguments{
\item{id}{ID number for the ASL sign}
}
\value{
The unique ID gloss linked to the sign used as its label in Signbank
}
\description{
This function inputs an ASL ID number as an integer or string and outputs the
unique ID gloss linked to the same sign (its Signbank entry).
}
