% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id2gloss_sts.R
\name{id2gloss_sts}
\alias{id2gloss_sts}
\title{Find ID gloss from ID number (Svenskt teckenspråkslexikon)}
\usage{
id2gloss_sts(id)
}
\arguments{
\item{id}{ID number for the STS sign}
}
\value{
The unique ID gloss linked to the sign used as its label in Svenskt teckenspråkslexikon
}
\description{
This function inputs an STS ID number as an integer or string and outputs the
gloss linked to the same sign (its entry in Svenskt teckenspråkslexikon).
}
