% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNotNumeric.R
\name{isNotNumeric}
\alias{isNotNumeric}
\title{Check if input is not numeric}
\usage{
isNotNumeric(x)
}
\arguments{
\item{x}{Any input value}
}
\value{
TRUE/FALSE
}
\description{
This function inputs value and returns TRUE if not numeric
}
