% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNumeric.R
\name{isNumeric}
\alias{isNumeric}
\title{Check if input is numeric}
\usage{
isNumeric(x)
}
\arguments{
\item{x}{Any input value}
}
\value{
TRUE/FALSE
}
\description{
This function inputs value and returns TRUE if numeric
}
