% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_avatar.R
\name{make_avatar}
\alias{make_avatar}
\title{Animate OpenPose signer}
\usage{
make_avatar(
  data,
  destination = "./signglossR_avatar.gif",
  skin_col = "lightgrey",
  clothes_col = "lightgrey",
  turtleneck = FALSE,
  fancy = FALSE
)
}
\arguments{
\item{data}{The data frame with OpenPose data in wide format}

\item{destination}{The path/filename of the output .gif}

\item{skin_col}{Skin color of the avatar}

\item{clothes_col}{Color of the avatar's clothes}

\item{turtleneck}{Whether the avatar should wear a turtleneck sweater (default=FALSE)}

\item{fancy}{Whether the avatar should be dressed in a tuxedo, top hat, and monocle (default=FALSE)}
}
\value{
A .gif file with an animated avatar
}
\description{
This function inputs a data frame from `read_openpose(wide=TRUE)` (.eaf) and outputs an animated .gif of an avatar signing
}
