% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_elan_image.R
\name{make_elan_image}
\alias{make_elan_image}
\title{Make ELAN image}
\usage{
make_elan_image(
  path,
  destination = "./",
  segmentation_tier = "",
  gloss_tier = "",
  video = "",
  combine = TRUE,
  scale = 1,
  crop = FALSE,
  region = "100x150+100+100",
  text = "",
  textcolor = "white",
  font = "Helvetica",
  fontsize = "40",
  gravity = "southwest",
  location = "+20+20",
  border = FALSE
)
}
\arguments{
\item{path}{The path to a single ELAN .eaf file or a directory with .eaf files}

\item{destination}{The path/filename of the output image}

\item{segmentation_tier}{List of specific tiers with segmentation annotations to be included}

\item{gloss_tier}{List of specific tiers with gloss annotations to be included}

\item{video}{Specify a video path to be used (if not, first linked video per eaf is used)}

\item{combine}{Represent each sign by an overlay of the first and last frame of the sign (default=TRUE)}

\item{scale}{The resizing proportion of the original image (0 to 1; default is 1 = original size)}

\item{crop}{Whether to crop image to region (`crop=FALSE` is default)}

\item{region}{The region to be cropped (defaults to '100x150+100+100', (width x height +upper_x +upper_y)). Only for `crop=TRUE`}

\item{text}{The text string to be added to the image}

\item{textcolor}{The color of the text annotation (default is "white")}

\item{font}{The font of the text annotation (default is "Helvetica")}

\item{fontsize}{The fontsize of the text annotation (default is 70)}

\item{gravity}{The anchor point of the added text (default is "southwest")}

\item{location}{The location of the text relative to the anchor point}

\item{border}{Set to `TRUE` if you want a black border frame for the output image}
}
\value{
The name of the image file(s) created
}
\description{
This function creates image files (with or without text) from an ELAN annotation file
}
