% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_elan_video.R
\name{make_elan_video}
\alias{make_elan_video}
\title{Make ELAN video}
\usage{
make_elan_video(
  path,
  destination = "./",
  segmentation_tier = "",
  gloss_tier1 = "",
  gloss_tier2 = "",
  gloss_gravity = "northwest",
  video = "",
  segment = TRUE,
  gloss_size = 1,
  subtitle_size = 1,
  padding = 0
)
}
\arguments{
\item{path}{The path to a single ELAN .eaf}

\item{destination}{The path/file of the output video}

\item{segmentation_tier}{List of specific tiers with segmentation annotations to be included}

\item{gloss_tier1}{Name of gloss tier with gloss annotations to be included}

\item{gloss_tier2}{Name of second gloss tier with gloss annotations to be included}

\item{gloss_gravity}{The gravity (=anchor point) of the text of the gloss text}

\item{video}{Specify a video path to be used (if not, first linked video is used)}

\item{segment}{Whether to cut each segmentation into a separate video (default=FALSE)}

\item{gloss_size}{Fontsize of gloss text in video (default=1; 0 results in no glosses)}

\item{subtitle_size}{Fontsize of subtitle text based on segmentation tier cell text in video (default=1; 0 results in no subtitles)}

\item{padding}{Whether to add n frames before and after the start/end to segment (default=0)}
}
\value{
A video file based on ELAN annotations
}
\description{
This function creates video file clips (with or without text) from an ELAN annotation file
}
