% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gif.R
\name{make_gif}
\alias{make_gif}
\title{Make a GIF from a videofile (with ffmpeg)}
\usage{
make_gif(file, fps = 12.5, scale = 0.5)
}
\arguments{
\item{file}{The path to the video to be converted}

\item{fps}{Set frame rate (fps: frames per second) for gif output (default is 12.5)}

\item{scale}{Scale the output video dimension to a factor (0 to 1) of the original (default is .5)}
}
\value{
The name of the output .gif file
}
\description{
This function inputs a video and converts it to .gif using ffmpeg and ImageMagick
}
