% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_image_ex.R
\name{make_image_ex}
\alias{make_image_ex}
\title{Make image example from an image file}
\usage{
make_image_ex(
  file,
  destfile = "",
  crop = FALSE,
  scale = 1,
  region = "100x150+100+100",
  text = "",
  textcolor = "white",
  font = "Helvetica",
  fontsize = "70",
  gravity = "southwest",
  location = "+20+20",
  border = FALSE
)
}
\arguments{
\item{file}{The path to the image to be modified}

\item{destfile}{The specified path to the modified file created}

\item{crop}{Whether to use automatic face detection or not (`automatic=TRUE` is default)}

\item{scale}{The resizing proportion of the original image (0 to 1; default is 1 = original size)}

\item{region}{The region to be cropped (defaults to '100x150+100+100', (width x height +upper_x +upper_y)). Only for `crop=TRUE`}

\item{text}{The text string to be added to the image}

\item{textcolor}{The color of the text annotation (default is "white")}

\item{font}{The font of the text annotation (default is "Helvetica")}

\item{fontsize}{The fontsize of the text annotation (default is 70)}

\item{gravity}{The anchor point of the added text (default is "southwest")}

\item{location}{The location of the text relative to the anchor point}

\item{border}{Set to `TRUE` if you want a black border frame for the output image}
}
\value{
The name of the image file that was created
}
\description{
This function inputs an image file and allows for resizing, cropping, and text annotations
}
