% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_video_ex.R
\name{make_video_ex}
\alias{make_video_ex}
\title{Make a video example from a videofile (with ffmpeg)}
\usage{
make_video_ex(video, destination = "./", speed = 0.5, rep = FALSE)
}
\arguments{
\item{video}{The path/filename to the video to be modified}

\item{destination}{The path/filename of the output video}

\item{speed}{Set the speed of the video to a factor (0 to 1) of the original (default is .5)}

\item{rep}{If output video is to be repeated through concatenation (default is `FALSE`)}
}
\value{
The name of the output video file
}
\description{
This function inputs a video and allows for repeated and/or slowmotion playback
}
