% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notNumeric.R
\name{notNumeric}
\alias{notNumeric}
\title{Check if input is not numeric}
\usage{
notNumeric(x)
}
\arguments{
\item{x}{Any input value}
}
\value{
TRUE/FALSE
}
\description{
This function inputs value and returns TRUE if not numeric
}
