% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_openpose.R
\name{plot_openpose}
\alias{plot_openpose}
\title{Plot OpenPose}
\usage{
plot_openpose(
  data,
  point_size = 5,
  hands = "both",
  right_color = "green3",
  left_color = "red3",
  path_color = FALSE,
  silhouette = 0
)
}
\arguments{
\item{data}{The data frame with OpenPose data in wide format}

\item{point_size}{The size of the hand location points (default=5)}

\item{hands}{Which hands to plot: options are `both` (default), `right`, and `left`}

\item{right_color}{The color of the right hand (keypoint 4) articulation (default="green")}

\item{left_color}{The color of the left hand (keypoint 7) articulation (default="red")}

\item{path_color}{Whether to color the articulation based on the movement path, chronologically lighter to darker (default=FALSE)}

\item{silhouette}{Whether to include a "ghost" silhouette of a signer as reference (default=0; 1 is fully visible)}
}
\value{
A ggplot2 object
}
\description{
This function inputs a data frame from `read_openpose(wide=FALSE)` (.eaf) and plots the positions of the hands (keypoints 4 and 7)
}
