% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_openpose.R
\name{read_openpose}
\alias{read_openpose}
\title{Read OpenPose data (.json)}
\usage{
read_openpose(path, wide = TRUE, fill_zero = FALSE)
}
\arguments{
\item{path}{The path to the directory with OpenPose JSON data}

\item{wide}{Whether the output data frame should be transformed to wide format (default=TRUE)}

\item{fill_zero}{Whether zero values should be filled with nearest non-zero value (default=FALSE)}
}
\value{
A data frame containing the OpenPose data
}
\description{
This function reads OpenPose generated JSON files in a directory and outputs a data frame
}
