% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_corpus.R
\name{search_corpus}
\alias{search_corpus}
\title{Search corpus}
\usage{
search_corpus(gloss = "*", acronym = "sts")
}
\arguments{
\item{gloss}{ID number for the sign}

\item{acronym}{The acronym for the target sign language (e.g. "STS")}
}
\value{
The url of the linked corpus search
}
\description{
This function inputs search string and searches it in an online corpus
}
