% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_elan_image.R
\name{segment_elan_image}
\alias{segment_elan_image}
\title{Segment image(s) from video based on ELAN annotations}
\usage{
segment_elan_image(
  eaf,
  segmentation_tier = NULL,
  gloss_tier = NULL,
  target_glosses = NULL,
  frame_offset = 0,
  gloss_text = F,
  video = NULL,
  image_format = ".jpg",
  path = NULL,
  tag = T,
  trim = 1,
  scale = 1,
  crop_region = NULL,
  text = NULL,
  textcolor = "white",
  boxcolor = NULL,
  font = "Helvetica",
  fontsize = "70",
  rel_fontsize = NULL,
  gravity = "southwest",
  location = "+20+20",
  border = FALSE
)
}
\arguments{
\item{eaf}{The ELAN files that supplies annotations as segments}

\item{segmentation_tier}{Specify tier (TIER_ID) for segments}

\item{gloss_tier}{Specify tier(s) (TIER_ID) for glosses}

\item{target_glosses}{Specify which glosses to be kept (default is all glosses)}

\item{frame_offset}{Select image +/-n from the gloss annotation's start frame (default = 0)}

\item{gloss_text}{Whether to include text from `gloss_tier` (default = `FALSE`)}

\item{video}{The filename of the video to be modified}

\item{image_format}{The format of the output images (default = '.jpg')}

\item{path}{The path to the output images}

\item{tag}{Whether to use annotations as tags in filenames (default = `TRUE`)}

\item{trim}{Optional argument for horizontal image trim (0 to 1; default = 1)}

\item{scale}{The resizing proportion of the original image (0 to 1; default is 1 = original size)}

\item{crop_region}{The region to be cropped (e.g. '100x150+100+100', (width x height +upper_x +upper_y)).}

\item{text}{The text string to be added to the image}

\item{textcolor}{The color of the text annotation (default is "white")}

\item{boxcolor}{The color of the textbox (default is `NULL`)}

\item{font}{The font of the text annotation (default is "Helvetica")}

\item{fontsize}{The fontsize of the text annotation (default is 70)}

\item{rel_fontsize}{Whether to adjust the fontsize relative to width (x) or height (y)}

\item{gravity}{The anchor point of the added text (default is "southwest")}

\item{location}{The location of the text relative to the anchor point}

\item{border}{Set to `TRUE` if you want a black border frame for the output image}
}
\value{
The names of the output image files
}
\description{
This function inputs an ELAN file and a video and
outputs images from the ELAN annotations as segments
}
