% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_elan_video.R
\name{segment_elan_video}
\alias{segment_elan_video}
\title{Segment a video file based on ELAN annotations}
\usage{
segment_elan_video(
  eaf,
  segmentation_tier = NULL,
  target_glosses = NULL,
  tag = T,
  video = NULL,
  video_format = ".mp4",
  path = NULL,
  method = "av",
  preset = "medium",
  crf = "12",
  padding = 0
)
}
\arguments{
\item{eaf}{The ELAN files that supplies annotations as segments}

\item{segmentation_tier}{Specify tier(s) (TIER_ID) to be read}

\item{target_glosses}{Specify which glosses to be kept}

\item{tag}{Whether to use annotations as tags in filenames (default = `TRUE`)}

\item{video}{The filename of the video to be modified}

\item{video_format}{The format of the output videos (default = '.mp4')}

\item{path}{The path to the output videos}

\item{method}{Whether to use `av` (in R) or `ffmpeg` (system call) (default = "av")}

\item{preset}{Video quality preset (default = "medium")}

\item{crf}{Video processing preset (default = "12")}

\item{padding}{Add (+) or subtract (-) padding at start/end (default = 0 msec)}
}
\value{
The names of the output video files
}
\description{
This function inputs an ELAN file and a video and
outputs new videos from the ELAN annotations as segments
}
