% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_elan_video.R
\name{split_elan_video}
\alias{split_elan_video}
\title{Split ELAN video}
\usage{
split_elan_video(
  elan_path = "",
  segmentation_tier = "",
  video_path = "",
  annotation_tag = F,
  padding = 0,
  video_input_format = ".mp4",
  video_output_format = ".mp4"
)
}
\arguments{
\item{elan_path}{The path to a single ELAN .eaf file or a directory with .eaf files}

\item{segmentation_tier}{List of specific tiers with segmentation annotations to be included}

\item{video_path}{Specify a video path to be used (if none, uses the ELAN path and filename)}

\item{annotation_tag}{If TRUE, uses the annotation text of the cell as the filename}

\item{padding}{Adds (or subtracts if negative) frames (in milliseconds) before+after segment duration}

\item{video_input_format}{Specify input video format in directory (default is .mp4)}

\item{video_output_format}{Specify output video format in directory (default is .mp4)}
}
\value{
Video files based on ELAN segmentations
}
\description{
This function reads ELAN annotation file (.eaf) in a directory and outputs a video
for each annotation cell as the segmentation and can label the output file with the annotation text
}
