% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts_padding.R
\name{sts_padding}
\alias{sts_padding}
\title{Return zero padded STS ID number}
\usage{
sts_padding(id)
}
\arguments{
\item{id}{And STS ID number input}
}
\value{
Zero padded ID number
}
\description{
This function inputs a string and returns it with correct zero padding
}
