% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_image.R
\name{censor_image}
\alias{censor_image}
\title{Censor image using blurring or solid black}
\usage{
censor_image(file, region = "100x150+100+100", method = "blur")
}
\arguments{
\item{file}{The path to the image to be modified}

\item{region}{The region to be modified (defaults to '100x150+100+100', (width x height +upper_x +upper_y))}

\item{method}{The method of censoring: "blur" or "black" ("blur" is default)}
}
\value{
The name of the image file that was modified
}
\description{
This function inputs an image and applies blurring or solid black
censoring of a defined region. It returns a new file in the same directoy.
}
