% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_source.R
\name{cite_source}
\alias{cite_source}
\title{Cite source}
\usage{
cite_source(acronym = "signglossR")
}
\arguments{
\item{acronym}{A sign language acronym used by signglossR}
}
\value{
The preferred citation format for the language resource used
}
\description{
This function inputs a sign language acronym (e.g. "ASL") and
outputs the preferred citation for the original source used in signglossR.
}
