% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_images.R
\name{combine_images}
\alias{combine_images}
\title{Combine images}
\usage{
combine_images(
  images,
  filename = NULL,
  path = NULL,
  stack = FALSE,
  overlay = FALSE,
  blend = 80,
  skip = 0,
  trim = 1
)
}
\arguments{
\item{images}{A vector of image files to be combined}

\item{filename}{The filename of the combined output image}

\item{path}{The path to the combined output image}

\item{stack}{Optional argument if concatenated image wanted: default FALSE gives horizontal sequence; TRUE gives vertically stacked}

\item{overlay}{Optional argument if overlay image wanted (defaults to FALSE)}

\item{blend}{Level of blending between images (if overlay = `TRUE`)}

\item{skip}{Whether to skip one image before applying pairwise composite (if overlay = `TRUE`)}

\item{trim}{Optional argument if horizontal trim wanted (values 0 to 1; default = 1)}
}
\value{
The name of the output image file
}
\description{
This function inputs a list of image files and outputs a concatenated or overlay image
}
