% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_images.R
\name{combine_images}
\alias{combine_images}
\title{Combine images}
\usage{
combine_images(
  images,
  destination = "./",
  stack = FALSE,
  overlay = FALSE,
  trim = 1
)
}
\arguments{
\item{images}{A list of image files (vector or comma-separated string) to be combined}

\item{destination}{The path/filename of the combined output image}

\item{stack}{Optional argument if concatenated image wanted: default FALSE gives horizontal sequence; TRUE gives vertically stacked}

\item{overlay}{Optional argument if overlay image wanted (defaults to FALSE)}

\item{trim}{Optional argument if horizontal trim wanted (values 0 to 1; default=1)}
}
\value{
The name of the output image file
}
\description{
This function inputs a list of image files and outputs a concatenated or overlay image
}
