% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gif.R
\name{get_gif}
\alias{get_gif}
\title{Get sign gif}
\usage{
get_gif(id, destination = "./", acronym = "sts")
}
\arguments{
\item{id}{ID number for the sign}

\item{destination}{The path/filename of the downloaded image}

\item{acronym}{The acronym for the target sign language (e.g. "STS")}
}
\value{
The path of the image file that was downloaded
}
\description{
This function inputs an  ID number and downloads the
GIF of the corresponding sign entry in the selected language.
}
