% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image.R
\name{get_image}
\alias{get_image}
\title{Get sign image}
\usage{
get_image(
  id,
  filename = NULL,
  path = NULL,
  lang = "sts",
  overlay = FALSE,
  skip = 0,
  blend = 80,
  trim = 1
)
}
\arguments{
\item{id}{ID number for the sign}

\item{filename}{The filename of the downloaded image}

\item{path}{The path to the downloaded image}

\item{lang}{The target sign language (i.e. "STS" or "ASL")}

\item{overlay}{Optional argument if overlay image wanted (default = `FALSE`)}

\item{skip}{Whether to skip one image before applying pairwise composite (if overlay = `TRUE`)}

\item{blend}{Level of blending between images (if overlay = `TRUE`)}

\item{trim}{Optional argument for horizontal image trim (0 to 1; default = 1)}
}
\value{
The path of the image file that was downloaded
}
\description{
This function inputs an ID number and downloads the image
of the corresponding sign entry in the selected language.
}
\details{
Arguments available for STS only:
Options include creating side-by-side or overlay images (default = `FALSE`)
and whether a horizontal trim (0 to 1; default = 1) should be applied.
}
