% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video.R
\name{get_video}
\alias{get_video}
\title{Get sign video}
\usage{
get_video(id, filename = NULL, path = NULL, lang = "sts")
}
\arguments{
\item{id}{ID number for the sign}

\item{filename}{The filename of the downloaded video}

\item{path}{The path to the downloaded video}

\item{lang}{The target sign language (i.e. "STS" or "ASL")}
}
\value{
The path of the video file that was downloaded
}
\description{
This function inputs an ID number and downloads the video
of the corresponding sign entry in the selected language.
}
