% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss2id.R
\name{gloss2id}
\alias{gloss2id}
\title{Find ID number from ID gloss}
\usage{
gloss2id(gloss, lang = "sts")
}
\arguments{
\item{gloss}{ID gloss for the sign}

\item{lang}{The target sign language ("STS" or "ASL")}
}
\value{
The unique ID gloss(es) linked to the gloss
}
\description{
This function inputs an ID gloss and outputs an
ID number linked to the same sign (its dictionary entry).
}
