% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id2gloss.R
\name{id2gloss}
\alias{id2gloss}
\title{Find ID gloss from ID number}
\usage{
id2gloss(id, lang = "sts")
}
\arguments{
\item{id}{ID number for the sign}

\item{lang}{The target sign language ("STS" or "ASL")}
}
\value{
The unique ID gloss linked to the sign in the dictionary
}
\description{
This function inputs an ID number as an integer or string and outputs the
unique ID gloss linked to the same sign (its dictionary entry).
}
