% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_elan.R
\name{read_elan}
\alias{read_elan}
\title{Reads an ELAN (.eaf) file}
\usage{
read_elan(eaf_file = NULL, tiers = NULL, target_annotations = NULL)
}
\arguments{
\item{eaf_file}{Path to ELAN (.eaf) file}

\item{tiers}{Specify tiers (TIER_ID or TIER_TYPE) to be read}

\item{target_annotations}{Specify which annotations to be kept (default `NULL` = all)}
}
\value{
Tibble of the ELAN annotations
}
\description{
Reads an ELAN (.eaf) file
}
