% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_elan.R
\name{read_elan}
\alias{read_elan}
\title{Read ELAN data (.eaf)}
\usage{
read_elan(path, segmentation_tier = "", gloss_tier = "", video = "")
}
\arguments{
\item{path}{The path to a single ELAN .eaf file or a directory with .eaf files}

\item{segmentation_tier}{List of specific tiers with segmentation annotations to be included}

\item{gloss_tier}{List of specific tiers with gloss annotations to be included}

\item{video}{Specify a video path to be used (if not, first linked video per eaf is used)}
}
\value{
A data frame containing the ELAN data
}
\description{
This function reads ELAN annotation file (.eaf) in a directory and outputs a data frame
}
